trying to fix screencomponent with 100vw and 100vh
