+++ src/components/beagle-screen/beagle-screen.component.html
@@ -1,4 +1,4 @@
-/*
+<!-- 
   * Copyright 2020 ZUP IT SERVICOS EM TECNOLOGIA E INOVACAO SA
   *
   * Licensed under the Apache License, Version 2.0 (the "License");
@@ -11,12 +11,7 @@
   * distributed under the License is distributed on an "AS IS" BASIS,
   * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
   * See the License for the specific language governing permissions and
-  * limitations under the License.
-*/
+  * limitations under the License. 
+-->
 
-import * as CSS from 'csstype'
-
-export interface BeagleContainerInterface {
-  onInit?: () => void,
-  style?: CSS.Properties,
-}
+<ng-content></ng-content>
