+++ src/components/index.ts
@@ -35,3 +35,4 @@ export * from './beagle-text-area/beagle-text-area.component'
 export * from './beagle-markdown/beagle-markdown.component'
 export * from './beagle-webview/beagle-webview.component'
 export * from './beagle-tab-bar/beagle-tab-bar.component'
+export * from './beagle-screen/beagle-screen.component'
