+++ src/components/module.ts
@@ -41,6 +41,7 @@ import {
 } from './beagle-page-indicator/beagle-page-indicator.component'
 import { BeagleWebviewComponent } from './beagle-webview/beagle-webview.component'
 import { BeagleTabBarComponent } from './beagle-tab-bar/beagle-tab-bar.component'
+import { BeagleScreenComponent } from './beagle-screen/beagle-screen.component'
 
 const components = [
   BeagleButtonComponent,
@@ -64,6 +65,7 @@ const components = [
   BeagleMarkdownComponent,
   BeagleWebviewComponent,
   BeagleTabBarComponent,
+  BeagleScreenComponent,
 ]
 
 @NgModule({
